quick refactor of the dashboard component to reuse the hooks at /dashboard
